Increase cookie lifetime to 400 days
Modern browsers don't allow cookie expiry to be set more than 400 days in the future - may as well use this for storing settings.  It is possible that older browsers such as found on eReaders may allow longer, so it may be good to set it much longer - newer browsers will just cap it at 400.